Migrate C testing to GoogleTest
