+++ .github/workflows/spglib-pytest-conda.yml
@@ -67,9 +67,12 @@ jobs:
         conda install --yes -c conda-forge gcc_linux-64 make cmake
     # Require CMake>=3.14 for the following build procedure
     - name: Test
+      working-directory: ./test
       run: |
         cmake -S . -B build -DUSE_SANITIZER="Address"
-        cmake --build build -t check
+        cmake --build build
+        cd build
+        ctest
 
   docs:
     runs-on: ubuntu-latest
